Refactor the python persister
Separate out the persister classes into their own python files
and clean up some unused code for start up.
