+++ monasca_persister/README.md
@@ -2,20 +2,20 @@
 
 A Monasca Persister written in Python.
 
-Reads alarms and metrics from a Kafka queue and stores them in an InfluxDB 
+Reads alarms and metrics from a Kafka queue and stores them in an InfluxDB
 database.
 
 
 # License
 
-Copyright (c) 2014 Hewlett-Packard Development Company, L.P.
+(C) Copyright 2014-2016 Hewlett Packard Enterprise Development Company LP
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
     http://www.apache.org/licenses/LICENSE-2.0
-    
+
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
