+++ monasca_persister/persister.conf
@@ -1,17 +1,18 @@
 [DEFAULT]
 log_file = persister.log
-log_dir = .
+log_dir = /var/log/monasca/persister
 # Default log level is WARNING
 # Show debugging output in logs (sets DEBUG log level output)
 debug = false
 # Show more verbose log output (sets INFO log level output) if debug is False
 verbose = true
 
-[database]
-# Choose database type
-# database_type := 'cassandra' | 'influxdb'
-#database_type = cassandra
-database_type = influxdb
+[repositories]
+# The driver to use for the metrics repository
+metrics_driver = monasca_persister.repositories.influxdb.metrics_repository:MetricInfluxdbRepository
+
+# The driver to use for the alarm state history repository
+alarm_state_history_driver = monasca_persister.repositories.influxdb.alarm_state_history_repository:AlarmStateHistInfluxdbRepository
 
 [zookeeper]
 # Comma separated list of host:port
@@ -59,7 +60,7 @@ port = 8086
 user = mon_persister
 password = password
 
-# Uncomment and set cluster_ip_addresses if database_type is 'cassandra'
+# Uncomment, set cluster_ip_addresses, and change the repositories to point to the cassandra classes
 #[cassandra]
 # Comma separated list of Cassandra node IP addresses. No spaces.
 #cluster_ip_addresses: 10.10.10.3
