+++ monasca_persister/repositories/cassandra/abstract_repository.py
@@ -1,0 +1,37 @@
+# (C) Copyright 2016 Hewlett Packard Enterprise Development Company LP
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+# implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import abc
+from cassandra.cluster import Cluster
+from cassandra.query import BatchStatement
+from oslo_config import cfg
+import six
+
+from repositories.abstract_repository import AbstractRepository
+
+
+@six.add_metaclass(abc.ABCMeta)
+class AbstractCassandraRepository(AbstractRepository):
+
+    def __init__(self):
+        super(AbstractCassandraRepository, self).__init__()
+        self.conf = cfg.CONF
+
+        self._cassandra_cluster = Cluster(
+                self.conf.cassandra.cluster_ip_addresses.split(','))
+
+        self.cassandra_session = self._cassandra_cluster.connect(
+                self.conf.cassandra.keyspace)
+
+        self._batch_stmt = BatchStatement()
