+++ monasca_persister/repositories/influxdb/__init__.py
@@ -1,0 +1,25 @@
+# (C) Copyright 2016 Hewlett Packard Enterprise Development Company LP
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+# implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from oslo_config import cfg
+
+influxdb_opts = [cfg.StrOpt('database_name'),
+                 cfg.StrOpt('ip_address'),
+                 cfg.StrOpt('port'),
+                 cfg.StrOpt('user'),
+                 cfg.StrOpt('password')]
+
+influxdb_group = cfg.OptGroup(name='influxdb', title='influxdb')
+cfg.CONF.register_group(influxdb_group)
+cfg.CONF.register_opts(influxdb_opts, influxdb_group)
